# Back/BanderaGo/user-service/dtos/output/update_user_out.go
package output

import "time"

type UpdateUserOut struct {
	ID        uint      `json:"id"`
	Name      string    `json:"name"`
	LastName  string    `json:"last_name"`
	UpdatedAt time.Time `json:"updated_at"`
}